[level 0] Title: 구슬을 나누는 경우의 수, Time: 0.30 ms, Memory: 31.2 MB -BaekjoonHub

+++ 프로그래머스/0/120840. 구슬을 나누는 경우의 수/구슬을 나누는 경우의 수.cs
@@ -1,18 +1,18 @@
 using System;
 
 public class Solution {
-    public double solution(int balls, int share) {
+    public long solution(int balls, int share) {
         return Comb(balls,share);
     }
-    public double Comb(int n, int r)
+    
+    public long Comb(int n, int r)
     {
-        if(r > n-r)r = n-r;
-        double result = 1;
-        for(int i = 1; i <= r; i++)
+        if(r > n-r) r = n-r;
+        long result =1;
+        for(int i = 1;i<=r; i++)
         {
             result *= n-i+1;
             result /= i;
-            
         }
         return result;
     }
